+++ src/main/java/gov/usgs/traveltimeservice/TravelTimeControllerClient.java
@@ -1,8 +1,8 @@
 package gov.usgs.traveltimeservice;
 
+import gov.usgs.processingformats.TravelTimeData;
 import gov.usgs.processingformats.TravelTimeException;
 import gov.usgs.processingformats.TravelTimeRequest;
-import gov.usgs.processingformats.TravelTimeData;
 import io.micronaut.http.annotation.Post;
 import io.micronaut.http.client.annotation.Client;
 import javax.validation.constraints.NotBlank;
@@ -11,5 +11,6 @@ import javax.validation.constraints.NotBlank;
 public interface TravelTimeControllerClient {
 
   @Post("/traveltime")
-  public TravelTimeData getLocation(@NotBlank TravelTimeRequest request) throws TravelTimeException;
+  public TravelTimeData getTravelTime(@NotBlank TravelTimeRequest request)
+      throws TravelTimeException;
 }
