+++ src/main/resources/application.properties
@@ -1,0 +1,11 @@
+micronaut.application.name=traveltimeservice
+
+micronaut.router.static-resources.swagger.enabled=true
+micronaut.router.static-resources.swagger.paths=classpath:META-INF/swagger
+micronaut.router.static-resources.swagger.mapping=/ws/traveltime/swagger/**
+
+micronaut.router.static-resources.public.enabled=true
+micronaut.router.static-resources.public.paths=classpath:public
+micronaut.router.static-resources.public.mapping=/**
+
+jackson.serialization.writeDatesAsTimestamps=false
