+++ src/main/resources/logback.xml
@@ -1,0 +1,16 @@
+<configuration>
+    <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
+        <!-- encoders are assigned the type
+             ch.qos.logback.classic.encoder.PatternLayoutEncoder by default -->
+        <encoder>
+            <pattern>%d{HH:mm:ss.SSS} [%thread] %-5level %logger{36} - %msg%n</pattern>
+        </encoder>
+    </appender>
+
+    <root level="info">
+        <appender-ref ref="STDOUT" />
+    </root>
+    <!--<logger name="io.micronaut.http" level="TRACE"/>-->
+    <!--<logger name="io.netty.handler.logging" level="TRACE" />-->
+    <!--<logger name="io.micronaut.jackson.parser" level="TRACE"/>-->
+</configuration>
