+++ src/main/resources/public/index.html
@@ -1,0 +1,12 @@
+<!doctype html>
+<html>
+  <head>
+    <meta http-equiv="refresh" content="0;URL='/ws/traveltime/'" />
+  </head>
+  <body>
+    <h1>TravelTime Web Service</h1>
+    <p>
+      <a href="/ws/traveltime/">Web service is at /ws/traveltime/</a>
+    </p>
+  </body>
+</html>
