+++ src/main/resources/public/ws/traveltime/index.html
@@ -1,0 +1,23 @@
+<!doctype html>
+<head>
+  <title>TravelTime Swagger Docs</title>
+  <meta charset="utf8"/>
+  <meta name="viewport" content="width=device-width, initial-scale=1"/>
+
+  <link rel="stylesheet" href="https://unpkg.com/swagger-ui-dist@3/swagger-ui.css"/>
+</head>
+<body>
+  <div id="swagger-ui"></div>
+
+  <script src="https://unpkg.com/swagger-ui-dist@3/swagger-ui-bundle.js"></script>
+  <script>
+    SwaggerUIBundle({
+        url: "./swagger/traveltime-service-0.0.yml",
+        dom_id: '#swagger-ui',
+        presets: [
+          SwaggerUIBundle.presets.apis
+        ]
+    })
+  </script>
+</body>
+</html>
